# specs/swaggerdoc.yaml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: 'Medten Product API'
  description: 'Interact with the Medten product database using this API.'
servers:
  - url: https://mgtwitxxi1.execute-api.ap-southeast-1.amazonaws.com/Prod/api/v1/
    description: API Test Environment
  - url: https://v73kddyeqk.execute-api.ap-southeast-1.amazonaws.com/Prod/api/v1/
    description: API POC (Proof of Concept) Environment
  - url: http://localhost:3000/api/v1/
    description: Developer Local Test Environments
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/gaumeister/medten-product-api/1.0.0
paths: 
...
  /groups:
    summary: Represents a product group
    post:
      summary: creates a group
      tags:
        - groups
      requestBody:
        description:  The group to be created.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
      responses:
        200:
          description: returns the group 
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Group'
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
  /groups/${id}:
    summary: Represents a product group.
    parameters:
    - name: id
      in: path
      description: the group id
      required: true
      schema:
        type: integer
    get:
      summary: gets group by id
      tags:
        - groups
      responses:
        200:
          description: OK
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Group'
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
    put:
      summary: updates a group
      tags:
        - groups
      requestBody:
        description: >
          The group to be updated.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Group'
      responses:
        200:
          description: returns the group 
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Group'
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
  /groups/search:
    post:
      summary: Perform an advanced search for groups.
      description: Extends basic search capabilities, allowing a search term to be used across many different group fields. Additional filter criteria can be used to further narrow the search-term search.
      tags:
        - groups
      requestBody:
        description: >
          The search criteria to use for the search.
        required: true
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/SearchCriteria'
              - type: object
                properties:
                  search_term:
                    type: string
                    description: wildcard search value to be used for searching any/all of the `search_term_fields`
                  search_term_fields:
                    description: which group fields the search term is applied to
                    type: array
                    items:
                      type: string
                      enum: [group_code]
                  
            examples:
              example1:
                description: >
                  Search for group having a `group_code` matching `028` (wildcard),
                  
                  
                  Order the results by `id` (group id). 
                  

                  Do not skip any results, but limit the output to `10` entries.
                value:
                  offset: 0
                  limit: 10
                  order_by:
                    - id
                  search_term: 028
                  search_term_fields:
                    - group_code
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    description: The total number of matching groups in the database.
                    type: integer
                  groups:
                    type: array
                    items:
                      $ref: '#/components/schemas/Group'
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
  /groups/search/download:
    post:
      summary: Download results from an advanced search for groups.
      description: Extends basic search capabilities, allowing a search term to be used across many different group fields. Additional filter criteria can be used to further narrow the search-term search.
      tags:
        - groups
      requestBody:
        description: >
          The search criteria to use for the search.
        required: true
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/SearchCriteria'
              - type: object
                properties:
                  search_term:
                    type: string
                    description: wildcard search value to be used for searching any/all of the `search_term_fields`
                  search_term_fields:
                    description: which group fields the search term is applied to
                    type: array
                    items:
                      type: string
                      enum: [group_code]
                  
            examples:
              example1:
                description: >
                  Search for group having a `group_code` matching `028` (wildcard),
                  
                  
                  Order the results by `id` (group id). 
                  

                  Do not skip any results, but limit the output to `10` entries.
                value:
                  offset: 0
                  limit: 10
                  order_by:
                    - id
                  search_term: 028
                  search_term_fields:
                    - group_code
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    description: The total number of matching groups in the database.
                    type: integer
                  groups:
                    type: array
                    items:
                      $ref: '#/components/schemas/Group'
                  headers:
                    type: array
                    items:
                      type: object
                      description: a single object whose properties can be used to create headers for CSV/table rendering.
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
components:
  schemas:
    ApiError:
      type: object
      properties:
        message:
          type: string
          description: A human-readable message describing the error.
        error:
          type: string
          description: The error details. Typically error details may contain sensitive information not intended to be shown directly in the user interface.
    Auditable:
      type: object
      properties:
        created:
          type: string
          format: date-time
          readOnly: true
        updated:
          type: string
          format: date-time
          readOnly: true
        version:
          type: integer
          readOnly: true
...
    Group:
      type: object
      description: The group used to group together families.
      allOf:
      - type: object
        properties:
          id:
            type: integer
            readOnly: true
          group_code:
            description: Code for the group
            type: string
            maxLength: 255
      - $ref: '#/components/schemas/Auditable'
...
    SearchCriteria:
      type: object
      allOf:
      - type: object
        properties:
          offset:
            type: integer
            description: for pagination. A zero-indexed value indicating where to begin the output with in the set of all possible search results.
          limit:
            type: integer
            description: for pagination. The number of records to return in a single query (up to 1000).
          order_by:
            type: array
            description: Describes the ordering of the search results. Provide the field names to order by. Prefix the field name with a `-` if you want ordering to be in descending order.
            items:
              type: string
          search_term:
            type: string
            description: the search term value. It will be treated as a wildcard search within the `search_term_fields` values.
          search_term_fields:
            type: array
            description: the list of fields you want the `search_term` to apply to. The criteria will be applied as a (OR) when conducting the search across more than one field.
            items:
              type: string        
...
  responses:
    400ClientError:
      description: client error
      content:
        application/json:
          schema: 
            $ref: '#/components/schemas/ApiError'
...
    500ServerError:
      description: server error
      content:
        application/json:
          schema: 
            $ref: '#/components/schemas/ApiError'